Split benchmarks.
